// src/enums/investorRelations/InvestorRelationEnum.ts
export enum InvestorRelationEnum {
  StockMarketInformation = 'stock-market-information',
  FinancialReports = 'financial-reports',
  Shareholder = 'shareholder',
  DividendPayment = 'dividend-payment',
  ShareholdersMeeting = 'shareholders-meeting',
  CompanyIntroductionVideo = 'company-introduction-video',
}